+++ modules/protocol/src/main/java/com/icodesoftware/TokenValidator.java
@@ -14,12 +14,12 @@ public interface TokenValidator {
     OnSession validatePassword(Access hash, String password);
 
     //generate the ticket with limited life circle
-    String ticket(long id,int stub);
+    String ticket(long id,long stub);
     //validate ticket, call before application callback
     boolean validateTicket(Session session);
 
 
-    void offSession(long id,int stub);
+    void offSession(long id,long stub);
 
 
     boolean validateToken(Map<String,Object> params);
