+++ platform/src/main/java/com/tarantula/game/SimpleStub.java
@@ -3,7 +3,7 @@ package com.tarantula.game;
 
 public class SimpleStub extends PlayerGameObject{
 
-    public SimpleStub(String systemId,int stub){
+    public SimpleStub(String systemId,long stub){
         this.systemId = systemId;
         this.stub = stub;
     }
