+++ platform/src/main/java/com/tarantula/game/StubKey.java
@@ -9,15 +9,15 @@ public class StubKey implements Recoverable.Key {
 
     private String label;
 
-    private int stub;
-    public StubKey(String systemId,String label,int  stub){
+    private long stub;
+    public StubKey(String systemId,String label,long  stub){
         String[] query = systemId.split("/");
         this.bucket = query[0];
         this.oid = query[1];
         this.label = label;
         this.stub = stub;
     }
-    public StubKey(String bucket, String oid,String label,int  stub){
+    public StubKey(String bucket, String oid,String label,long  stub){
         this.bucket = bucket;
         this.oid = oid;
         this.label = label;
