+++ platform/src/main/java/com/tarantula/platform/event/PermissionCheckEvent.java
@@ -22,7 +22,7 @@ public class PermissionCheckEvent extends Data implements EventOnAction {
 		out.writeUTF("2",this.sessionId);
         out.writeLong("3",this.distributionId);
 		out.writeUTF("4",this.token);
-        out.writeInt("5", this.stub);
+        out.writeLong("5", this.stub);
 		out.writeUTF("6",this.action);
 		out.writeUTF("7",this.ticket);
 		out.writeInt("8",this.routingNumber);
@@ -34,7 +34,7 @@ public class PermissionCheckEvent extends Data implements EventOnAction {
 		this.sessionId = in.readUTF("2");
         this.distributionId = in.readLong("3");
 		this.token = in.readUTF("4");
-		this.stub = in.readInt("5");
+		this.stub = in.readLong("5");
 		this.action = in.readUTF("6");
 		this.ticket = in.readUTF("7");
 		this.routingNumber = in.readInt("8");
